elesma_project/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

# Put project-specific requirements here.
# See http://pip-installer.org/requirement-format.html for more information.

django-haystack==1.0.1-final
Whoosh==1.0.0b4
django-ratings==0.3.4
django-taggit==0.8.0
